+++ src/Generador/Generador.h
@@ -12,5 +12,5 @@ namespace Generador{
         ALFABETICA_MINUSCULA = 6,
         NUMERICA = 7
     };
-    unsigned char* generarContrasenia(const uint8_t& CANT_CARACTERES, const TipoContrasenia& TIPO);
+    unsigned char* generarContrasenia(const unsigned int& CANT_CARACTERES, const TipoContrasenia& TIPO);
 }
